implemented frustum culling, did stupid bin+lib hack so rustdoc will run doctests properly
