+++ src/world/dimension.rs
@@ -1,6 +1,5 @@
 //! A dimension.
 
-
 use std::sync::{Arc, RwLock};
 use std::sync::atomic::AtomicUsize;
 
@@ -20,6 +19,14 @@ pub struct Dimension {
 }
 
 
+/// Radius in meters around the player to load chunks.
+const CHUNK_DISTANCE: f32 = 128.0;
+/// Radius in chunks.
+const CHUNK_RADIUS: i32 = (CHUNK_DISTANCE / ::world::CHUNK_SIZE_F32) as i32;
+/// Offset added to distance before removing chunks. (Must be > 0 to prevent load/unload loops)
+const UNLOAD_OFFSET: f32 = 8.0;
+
+
 impl Dimension {
     pub fn new() -> Dimension {
         Dimension {
@@ -30,8 +37,6 @@ impl Dimension {
 
     /// Adds new chunks as the player moves closer to them.
     pub fn load_chunks(&mut self, player_pos: Point3<f32>, queue_arc: Arc<RwLock<RenderQueue>>) {
-        const CHUNK_RADIUS: i32 = 3;
-
         let gen = ::world::generators::PerlinGenerator::new(0); // TODO: use seed
         let player_x_in_chunks = (player_pos.x / CHUNK_SIZE_F32) as i32;
         let player_y_in_chunks = (player_pos.y / CHUNK_SIZE_F32) as i32;
@@ -43,7 +48,13 @@ impl Dimension {
         for cx in (player_x_in_chunks-CHUNK_RADIUS)..(player_x_in_chunks+CHUNK_RADIUS+1) {
             for cy in (player_y_in_chunks - CHUNK_RADIUS)..(player_y_in_chunks + CHUNK_RADIUS + 1) {
                 for cz in (player_z_in_chunks - CHUNK_RADIUS)..(player_z_in_chunks + CHUNK_RADIUS + 1) {
-                    nearby_positions.push((cx as i32, cy as i32, cz as i32));
+                    let chunk_pos = (cx as i32, cy as i32, cz as i32);
+                    let chunk_world = Chunk::chunk_pos_to_center_ws(chunk_pos);
+                    let dist = Point3::distance(Point3::new(chunk_world.0, chunk_world.1, chunk_world.2), player_pos);
+                    if dist < CHUNK_DISTANCE {
+                        // only add chunks if they're close enough in a circular radius, not a square
+                        nearby_positions.push(chunk_pos);
+                    }
                 }
             }
         }
@@ -103,14 +114,12 @@ impl Dimension {
 
     /// Removes old chunks as the player moves away.
     pub fn unload_chunks(&mut self, player_pos: Point3<f32>, queue: Arc<RwLock<RenderQueue>>) {
-        const CHUNK_RADIUS: i32 = 3;
-        const CHUNK_DISTANCE: f32 = CHUNK_RADIUS as f32 * CHUNK_SIZE_F32;
         let mut chunks = self.chunks.write().unwrap();
         let old_num = chunks.len();
         chunks.retain(|pos, _| {
             let center = Chunk::chunk_pos_to_center_ws((pos.0, pos.1, pos.2));
             let dist = Point3::distance(Point3::new(center.0 as f32, center.1 as f32, center.2 as f32), player_pos);
-            dist < CHUNK_DISTANCE + 4.0 // offset added to prevent load/unload loop on the edge
+            dist < CHUNK_DISTANCE + UNLOAD_OFFSET // offset added to prevent load/unload loop on the edge
         });
         if chunks.len() != old_num {
             let mut lock = queue.write().unwrap();
