+++ src/world/generators/mod.rs
@@ -1,14 +1,29 @@
+//! World voxel data generators.
+
 use voxel::subdivmath::OctPos;
 use voxel::subdivstorage::{NaiveVoxelOctree, NaiveOctreeNode, SubdivVoxelDrain};
 use cgmath::{Vector3, MetricSpace};
 use noise::{NoiseFn, Perlin, Seedable};
 use world::{CHUNK_SIZE, CHUNK_SIZE_F32};
 
+
+/// Trait for world voxel generators.
 pub trait ChunkGenerator {
+    /// Generate a chunk.
+    ///
+    /// # Arguments
+    ///
+    /// * `pos` - Position in chunks.
+    ///
+    /// # Returns
+    ///
+    /// The octree generated for the given chunk.
     fn generate(&self, pos: (i32, i32, i32)) -> NaiveVoxelOctree<u8, ()>;
 }
 
+
 #[allow(dead_code)]
+/// Test generator. Generates a sphere in every chunk.
 pub struct SphereGenerator {
     seed: u32,
 }
@@ -27,14 +42,14 @@ impl ChunkGenerator for SphereGenerator {
     fn generate(&self, _pos: (i32, i32, i32)) -> NaiveVoxelOctree<u8, ()> {
         let mut tree = NaiveVoxelOctree{scale : 6 , root: NaiveOctreeNode::new_leaf(0)};
 
-        let center = Vector3::new(CHUNK_SIZE_F32, CHUNK_SIZE_F32, CHUNK_SIZE_F32);
+        let center = Vector3::new(CHUNK_SIZE_F32 / 2.0, CHUNK_SIZE_F32 / 2.0, CHUNK_SIZE_F32 / 2.0);
 
         for x in 0..CHUNK_SIZE {
             for y in 0..CHUNK_SIZE {
                 for z in 0..CHUNK_SIZE {
                     let pos = Vector3::new(x as f32, y as f32, z as f32);
                     let dist = pos.distance(center);
-                    if dist < 15.0 {
+                    if dist < 10.0 {
                         tree.set(opos!((x, y, z) @ 0), 1).unwrap();
                     }
                 }
@@ -45,15 +60,27 @@ impl ChunkGenerator for SphereGenerator {
     }
 }
 
+
+/// Simple perlin noise terrain generator.
 pub struct PerlinGenerator {
+    /// perlin noise generator
     perlin: Perlin,
+    /// noise scale
     scale: f64,
+    /// noise offset
     offset: f64,
+    /// secondary perlin generator for block types
     block_type_noise: Perlin,
+    /// scale for block type noise
     block_type_scale: f64,
 }
 
 impl PerlinGenerator {
+    /// Creates a new PerlinGenerator.
+    ///
+    /// # Arguments
+    ///
+    /// * `seed` - The random seed for this generator.
     pub fn new(seed: u32) -> PerlinGenerator {
         let perlin = Perlin::new();
         perlin.set_seed(seed);
@@ -72,6 +99,7 @@ impl PerlinGenerator {
 }
 
 impl ChunkGenerator for PerlinGenerator {
+    /// See [ChunkGenerator]
     fn generate(&self, pos: (i32, i32, i32)) -> NaiveVoxelOctree<u8, ()> {
         let mut tree = NaiveVoxelOctree{scale : 6 , root: NaiveOctreeNode::new_leaf(0)};
